[FE] Fix: Carousel.Text에서  Animate.Text 컴포넌트로 변경 및 수정
