+++ frontend/src/components/Animate/Text/index.tsx
